refactor: ♻️ move button customization out of full calendar component
